% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_package.R
\name{ajout_package}
\alias{ajout_package}
\title{Ajout d'un package a la liste}
\usage{
ajout_package(nom, package)
}
\arguments{
\item{nom}{nom de l'appli}

\item{package}{nom du package}
}
\description{
Ajout d'un package a la liste
}
