% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test.R
\name{create_test}
\alias{create_test}
\title{Creer un script pour tester le code}
\usage{
create_test(nom_script)
}
\arguments{
\item{nom_script}{nom du scrip a creer}
}
\value{
NA
}
\description{
Creer un script pour tester le code
}
