% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fct.R
\name{extract_fct}
\alias{extract_fct}
\title{Extrait les fonctions du script fonction.R en fichiers individuels}
\usage{
extract_fct(nom, version = NA)
}
\arguments{
\item{nom}{nom de l'appli}

\item{version}{nom de la version}
}
\value{
NA
}
\description{
Extrait les fonctions du script fonction.R en fichiers individuels
}
