% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst_pck.R
\name{inst_pck}
\alias{inst_pck}
\title{Creation d'un script pour installer les packages}
\usage{
inst_pck(nom)
}
\arguments{
\item{nom}{nom de l'appli}
}
\value{
NA
}
\description{
Creation d'un script pour installer les packages
}
