% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_prod.R
\name{launch_prod}
\alias{launch_prod}
\title{Lance l'appli en production}
\usage{
launch_prod(version)
}
\arguments{
\item{version}{version de l'appli}
}
\value{
NA
}
\description{
Lance l'appli en production
}
