% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_app.R
\name{new_app}
\alias{new_app}
\title{Creation d'une appli shiny}
\usage{
new_app(nom, trad = TRUE)
}
\arguments{
\item{nom}{nom de l'application shiny}

\item{trad}{booleen, si TRUE permet la création d'un fichier trad}
}
\value{
creer l'ensemble des fichiers et dossiers necessaires aux appli shiny
}
\description{
Creation d'une appli shiny
}
