% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fct.R
\name{new_fct}
\alias{new_fct}
\title{Créer une nouvelle fonction dans l'appli}
\usage{
new_fct(nom_app = NA, nom_fct)
}
\arguments{
\item{nom_app}{nom de l'application shiny}

\item{nom_fct}{nom de la fonction}
}
\value{
NA
}
\description{
Créer une nouvelle fonction dans l'appli
}
