% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Nouveau projet R
Création d'une arborescence pour structure les projets}
\usage{
new_project()
}
\value{
NA
}
\description{
Nouveau projet R
Création d'une arborescence pour structure les projets
}
