% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_subproject.R
\name{new_subproject}
\alias{new_subproject}
\title{Initialisation d'un sous projet}
\usage{
new_subproject(nom)
}
\arguments{
\item{nom}{nom du sous projet}
}
\value{
NA
}
\description{
Initialisation d'un sous projet
}
