% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_prod.R
\name{open_prod}
\alias{open_prod}
\title{Ouvre les scripts de l'appli en prod}
\usage{
open_prod(nom, version)
}
\arguments{
\item{nom}{nom de l'appli}

\item{version}{nom de la version}
}
\description{
Ouvre les scripts de l'appli en prod
}
