% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_script.R
\name{open_script}
\alias{open_script}
\title{Ouvre l'ensemble des scripts pour l'appli en cours (dev)}
\usage{
open_script(nom)
}
\arguments{
\item{nom}{nom de l'appli shiny}
}
\value{
NA
}
\description{
Ouvre l'ensemble des scripts pour l'appli en cours (dev)
}
