% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_trad_txt_output.R
\name{shiny_trad_txt_output}
\alias{shiny_trad_txt_output}
\title{Automatisation des traduction des textOutput}
\usage{
shiny_trad_txt_output(nom, extract = "new", type_sortie = "console")
}
\arguments{
\item{nom}{nom de l'appli shiny}

\item{extract}{new ou global, permet d'extraire les nouveaux textoutput ou l'ensemble}

\item{type_sortie}{console ou script, sortie dans la console R ou dans le script textoutput.R}
}
\value{
un tableau avec les noms des textOutput et les lignes de commandes à rentrer dans le fichier server
}
\description{
Automatisation des traduction des textOutput
}
