% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_prod.R
\name{to_prod}
\alias{to_prod}
\title{Archivage de l'appli en prod}
\usage{
to_prod(nom, version)
}
\arguments{
\item{nom}{nom de l'appli}

\item{version}{version de l'appli}
}
\value{
NA
}
\description{
Archivage de l'appli en prod
}
