% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_project.R
\name{view_project}
\alias{view_project}
\title{Ouvre les scripts d'un sous projet}
\usage{
view_project(nom)
}
\arguments{
\item{nom}{nom du sous projet}
}
\value{
NA
}
\description{
Ouvre les scripts d'un sous projet
}
